pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegisai"
version = "0.1.0"
description = "Local multimodal AI stack (enterprise privacy engine) — FastAPI control plane"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.0",
  "chromadb>=0.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=6.0.0",
  "ruff>=0.8.0",
  "build>=1.2.0",
]
otel = [
  "opentelemetry-api>=1.28.0",
  "opentelemetry-sdk>=1.28.0",
  "opentelemetry-instrumentation-fastapi>=0.49b0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/aegisai"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
